use prometheus directly for metrics
It occurred to me that we don't want to accumulate potentially
millions of per-site metrics forever, and that the `metrics`
crate doesn't provide a way to de-register metrics.

This commit bypasses that layer and goes directly to the prometheus
crate which does provide some nicer accessors that do allow
removing metrics.