shutdown if we have errors setting up the spool
Adds some plumbing that allows the application to request
a shutdown that is channelled along the same path as ctrl-c
handling.

If we have errors defining spools or during spool start,
then we request a shutdown on the basis that we can't store
anything and it would be unsafe to continue.