queue: fix priority inversion issue with QueueManager
QueueManager::insert could acquire MANAGER lock then a queue handle
lock.

Meanwhile, the queue maintainer would hold a queue handle lock
and then attempt to acquire MANAGER lock when idling out and removing
itself, leading to a livelock because MANAGER is protected by a
tokio Mutex which ping pongs in and out of the event loop.