include line table debug info in release builds
This gives a higher quality panic stack trace at the cost of inflating
the executable size from ~50MB -> ~500MB.

Full debug info is a possibility, but the executable size is ~800MB,
which is high, but not completely crazy. I don't know if we'll really
need that, so going with line debug for now.