docs: avoid copying $ prefixes from console code blocks
Rather than using `bash` for the mix of shell and output,
use `console`. That will cause the text to be marked up so
that the `$` prompt stuff won't get selected by dragging.

Enable copy button to copy the whole block, and add some
js to filter out the the $ and # prefixes from that content.