dns: fix ipv6 lookups
A couple of issues in here:

1. a copypasta incident meant that the v6 lookup was really just a v4
   lookup. doh!
2. We would lookup v4 addresses and just be satisfied with those,
   ignoring any v6 addresses

We now run the v4 and v6 lookups concurrently and merge the results
together.

We do still list the v4 addresses first in the list of addresses.
We don't try to be smart about using the "right" addresses when
work through the list; we try them all until one works out.

Correctly guessing at which one will work means replicating logic
in the kernel, and would potentially be screwed up when we add in
HAProxy support later.