docs: fixup changelog
Give the `main` page a non-empty title in the TOC.
The intent is that we'll roll that page over to a release specific
page when we cut a release and then empty out `main` again.

Enable the magiclink extension for mkdocs, so that GH issue number
links work without any additional ceremony.