Fix Domain::Tagged to_string issue
When given an extension/tagged domain, we didn't include the closing
square bracket when rendering it as text.

Note that this looks like an IPv6 address, but it is not a conforming
IPv6 address which looks like `[IPv6:aaaa::bbbb]` rather than
`[aaaa:bbbb]`. The latter uses tag "aaaa" to define a different
addressing namespace that has no known implementation.

<https://datatracker.ietf.org/doc/html/rfc5321#section-4.1.3>