remove lifetime from the Signer struct
It is rather difficult to reuse the Signer in a dynamic application with
the lifetime present in the struct.

This commit removes the lifetime to make it easier to cache and reuse.

The logger has also been removed from the API for the signer. It was
used only in one place to log a single debug line, and it is a bit of an
imposition on the embedding application to require the use of the `slog`
crate just for that, and its removal helps with the goal of removing
lifetimes from the struct as well.

Some builder methods have been made a little more ergonomic by accepting
values that are convertible to String rather than requiring str
references.

Since this commit changes the API signature, it also bumps the version
from 0.2 to 0.3.