DNS resolution failures now increase num_attempts
Previously, if we encountered a DNS resolution error we'd reinsert
the associated message into the scheduled queue for retry in about
a minute.

That doesn't respect the retry interval or, as it turns out, the
queue expiration policy.

This commit adjusts the logic so that we do bump the number of
attempts and check the expiration.

This change will also apply to errors triggered by lua/webhook
code that fails to create a connection object, and for maildir
writes that fail.

closes: https://github.com/KumoCorp/kumomta/issues/65