dns: resolve ambiguity of null mx and failed host lookups
Tom "Edge Case" Mairs found an issue where what I think happened
was that the DNS records changed since we first established our
own internal data structures.

When the new set of records doesn't resolve to any A or MX entries
we'd return the same empty list of addresses that we used to signal
a null mx and incorrectly fail the remainder of that queue.

This commit introduces a separate state for the null mx state so that
we can instead delay the queue.