bug: we'd forget about a message when all connection attempts failed
Relatively serious: if none of the resolved hosts were reachable,
the message that initiated that attempt would get logged as a transfail
but we'd then forget about it until a restart when we'd re-discover
it from the spool.

This is resolved by not taking the message from the dispatcher and
instead leaving it in place so that its centralized Drop processing
will requeue it to the scheduled queue.