queue: centralize insert_ready logic
Commit f3be8727913298646c5ab125832ce961a7a0d4aa was only an issue
because we had essentially two copies of the logic to insert
into the ready queue.

This commit refactors the second location to work in terms of the
first.

In doing so, I think this also fixes an issue with queues configured
for non-SMTP delivery, such as lua delivery implementations for
webhooks/AMQP in the delayed->ready case.  I think this explains
some weirdness that Tom has seen in the past when playing with
webhook logging.