smtp client: ensure all reads and writes have a timeout
There were a few cases where client operations weren't wrapped
up in a timeout, which is a potential cause of hanging on shutdown:
if we're waiting an unbounded amount of time for a connection to
do something before we can shutdown, then we'll never finish.

The current timeouts struct really encodes the read timeout for
waiting for the result of a command.

What we also need is a duration to wait for sending data. We're
using the smaller of the per-command read timeout or 1 minute for
this.