fix NXDOMAIN issue showing up as no non-zero-weighted sources
The suspension logic changed things so that we'd trip over a DNS
resolution earlier than normal, inside the round robin routine.

That isn't capable of reporting the domain resolution issue and
it would bubble up as NoSources and we'd miscategorize the failure.

This commit addresses that by allowing the round robin to pick a
source and then we can subsequently bubble up and handle the domain
resolution issue as normal.