check_relaying: adjust precedence of relay_to vs relay_from
Eagle-eyed viewers of https://www.youtube.com/watch?v=Q-5E1Q5LlTY
may have noticed that my claim that things were working right at
the end were false! I'd misread the final SMTP status, but didn't
notice until later reviewing the stream.

In poking around at this a bit more, the issue is not due to the
changes made to support the new connection metadata, but rather
due to the precedence of the sender-based relaying options vs.
the receipient-based relaying restriction.  We'd always take
the `relay_to` setting as authoritative before considering the
sender based options, which made it rather difficult to get a
working relaying policy.

This commit flips that around: if the sender based rules permit
relaying, then they will take precedence.  Otherwise, we consider
the recipient restrictions.

I suspect that this is the root of the issue described in
https://github.com/KumoCorp/kumomta/issues/63 but for which I
wasn't able to reproduce with a simple local configuration.