ready_queue: disambiguate queue names by protocol
Include `@protocol-info` on the end of the ready queue names.

This prevents surprising false sharing of custom delivery protocols
when returning different protocols based on the tenant or campaign
metadata.

This is visible in the stats:

```json
  "total_connection_count": {
    "help": "total number of active connections ever made",
    "type": "counter",
    "value": {
      "service": {
        "smtp_client": 1.0,
        "smtp_client:source2->(in1-smtp|in2-smtp).messagingengine.com@smtp": 1.0
      }
    }
  },
```

and perhaps unexpectedly in the `site_name` field of the json log
records. That will need to get tidied up in a separate commit.