tsa: create schema to recall events and actions
This is using an in-memory sqlite db to keep track of events
and actions that we trigger.

In the future, the db will be persisted at a configurable location
on local storage.

Next step is to add endpoints for both config overrides and suspensions
that can be consumed by the mta nodes.