add egress_path_config suspend option, use that for tsa
The simplest way to convey suspension status is via a config override.
So let's allow configuring a path as suspended and deliver that
alongside the other configuration shared by the tsa-daemon.

Some additional work is needed in kumod to ensure that we reload
the path config while the ready queue is in existence; that will
occur in a follow up commit.