ready_queue: avoid potential counter race when draining ready queue
There were a few cases where we'd drain the fifo and then explicitly
set the metrics counter to 0.  Critically, the drain and the counter
update were not atomic wrt. other actors that might be inserting
data into the queue, so there was potential to race and perturb
the ready queue count.

This commit avoids every unilaterally setting the new value in
the ready queue module, and instead factors out the drain operation
to collect the messages and then adjust the count by the number
removed.