mailparsing: begin migrating to nom
I'd initially avoided this because I thought that pest would make
things easier overall, but I'm facing an impossible to debug
situation with the pest parser because the error/diagnostics
are not very friendly.

This commit introduces some plumbing that uses nom to build
the parser; the error reporting is much easier to understand
as we can annotate context information and display it in
an error-chain style with a helpful caret to highlight the
problem spots.

I'll convert more over in subsequent commits; this one is getting
large so this is a checkpoint.