catch kumo.on called inside kumo.on
Event handlers must be registered at the file scope and not from
within another event handler in order for events to be consistently
triggered and handled.

In particular, the `init` event is only ever triggered once on
server startup.  If other handlers are registered from within the
init event, those will only ever fire when we re-use that original
lua context, which is good for a limited number of uses before it
is aged out of the resource pool.  That might work under very
limited or lightweight testing scenarios, but will otherwise
result in very inconsistent behavior.

We now catch and prevent, with a very visible error, attempting
to call `kumo.on` from within an event handler.