mailparsing: support removing parts, headers and writing the result
This commit adds methods that will reconstitute a message from the
set of parsed headers and parts.

This mode doesn't normalize or fix up the content, it just emits
what it discovered during parsing.

The intent is to support the methods of the message crate that
add or remove headers.