http injection: rebuild full-content injections
This makes it super easy to template the whole message; it will be
parsed and rebuilt with appropriate transfer encoding and wrapping.

Fixup an issue where some of the header fields in the builder variant
of the API were not templated or added to the message.

Skip adding transfer encoding header when it is not needed.