disable Nagle for smtp
I was listening to this discussion (linked below) and realized that we
don't really need Nagling on SMTP and that we haven't explicitly
configured it.

I haven't measured the impact of this (shame on me?), but I'm not
anticipating it to have a noticeable impact in practice because SMTP is
very chatty already. I don't think it will hurt us either, as we're
careful to always write a full, actionable line at a time to the kernel,
so we shouldn't be penalized for small writes.

https://www.youtube.com/watch?v=mqvVmYhclAg