fix an issue with dot-stuffing in outgoing smtp
This was a 2-in-1 issue really.

When a message line starts with a `.` we are required to encode it
by prefixing the line with an additional `.`, aka: "dot stuffing".

In our implementation of this we would unintentionally remap CRLF
to CR, and then fail to emit CRLF.CRLF at the end of the data
that we'd planned to send.

The result of this was that we're hang waiting for the remote host
to respond to our DATA, meanwhile, it was waiting for the end of
our data.  We'd sit there until either side reached their respective
timeout for the data phase.

This commit fixes up both parts of this bug.